Evidence strength scoring
 - comes straight out of the list of evidence types, except...
 - ...when you're assessing 'custom' evidence, and you have to figure
   it out yourself, and we store that in a new field.
 - also had to re-organise the tests a bit because Ruby.
