+++ app/helpers/evidence_helper.rb
@@ -7,4 +7,13 @@ module EvidenceHelper
       evidence.evidence_type_other
     end
   end
+
+  def evidence_strength(form, evidence)
+    if evidence.evidence_type
+      item = form.lists['evidence_type'].items[evidence.evidence_type]
+      item.strength
+    else
+      evidence.custom_strength
+    end
+  end
 end
