+++ app/views/assessments/evidence/result.html.erb
@@ -2,18 +2,15 @@
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-l">Your score for evidence</h1>
     <p>
-      The <%= evidence_short_name(@form, @evidence) %> has a strength score of XX out of 4.
+      The <%= evidence_short_name(@form, @evidence) %> has a strength score of <%= evidence_strength(@form, @evidence) %> out of 4.
     </p>
 
     <p>
-      The way you check if the armed forces identity card is real has a validity score of YY out of 4.
+      The way you check if the <%= evidence_short_name(@form, @evidence) %>  card is real has a validity score of YY out of 4.
     </p>
 
     <p>
-      To find out why you got these scores, you can <a
-      href="https://www.gov.uk/government/publications/identity-proofing-and-verification-of-an-individual/identity-proofing-and-verification-of-an-individual#strength" target="_blank"
-      >read the guidance about identity proofing and verification <span class="new-tab">(link opens
-      in a new tab)</span></a>.
+      To find out why you got these scores, you can <a href="https://www.gov.uk/government/publications/identity-proofing-and-verification-of-an-individual/identity-proofing-and-verification-of-an-individual#strength" target="_blank">read the guidance about identity proofing and verification <span class="new-tab">(link opens in a new tab)</span></a>.
     </p>
 
     <%= form_tag("#{request.fullpath}", method: "post") do %>
