+++ app/views/assessments/overview.html.erb
@@ -62,8 +62,8 @@
             <tbody class="govuk-table__body">
               <tr class="govuk-table__row">
                 <td class="govuk-table__cell" scope="row">Strength</td>
-                <td class="govuk-table__cell" scope="row">
-                  ??
+                <td class="govuk-table__cell" scope="row" class="evidence-strength">
+                  <%= evidence_strength(@evidence_form, evidence) %>
                 </td>
               </tr>
               <tr class="govuk-table__row">
