+++ config/forms/evidence.yaml
@@ -128,6 +128,17 @@ lists:
         value: local_authority_doc_letter
         strength: 1
 
+  - name: custom_strength
+    items:
+      - text: '1'
+        value: 1
+      - text: '2'
+        value: 2
+      - text: '3'
+        value: 3
+      - text: '4'
+        value: 4
+
   - name: official_templates_update_frequency
     items:
       - text: Every year
