+++ db/migrate/20190628150714_add_custom_strength_to_evidence.rb
@@ -1,0 +1,5 @@
+class AddCustomStrengthToEvidence < ActiveRecord::Migration[5.2]
+  def change
+    add_column :evidence, :custom_strength, :int
+  end
+end
