+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_06_27_090600) do
+ActiveRecord::Schema.define(version: 2019_06_28_150714) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -92,6 +92,7 @@ ActiveRecord::Schema.define(version: 2019_06_27_090600) do
     t.json "uv_ir_features"
     t.json "crypto_features"
     t.boolean "is_complete", default: false, null: false
+    t.integer "custom_strength"
     t.index ["assessment_id"], name: "index_evidence_on_assessment_id"
   end
