+++ spec/system/add_evidence_flow_spec.rb
@@ -1,5 +1,6 @@
 require "rails_helper"
 require_relative "steps_helper"
+require_relative "evidence_steps_helper"
 
 RSpec.feature 'Add evidence flow', type: :system do
   scenario 'Add new evidence to an assessment (short version)' do
@@ -7,11 +8,11 @@ RSpec.feature 'Add evidence flow', type: :system do
     and_i_choose_a_regular_confidence_level
     and_i_add_a_new_piece_of_evidence('Passport or travel document', 'UK passport')
     and_i_answer_no_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('4', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('UK passport')
     and_i_add_a_new_piece_of_evidence('Certificate', 'Marriage or civil partnership certificate')
     and_i_answer_no_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('2', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('Marriage or civil partnership certificate')
   end
 
@@ -19,8 +20,9 @@ RSpec.feature 'Add evidence flow', type: :system do
     when_i_start_a_new_assessment
     and_i_choose_a_regular_confidence_level
     and_i_add_a_new_piece_of_other_evidence('Something else')
+    and_i_say_the_evidence_strength_is('3')
     and_i_answer_no_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('3', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('Something else')
   end
 
@@ -28,8 +30,9 @@ RSpec.feature 'Add evidence flow', type: :system do
     when_i_start_a_new_assessment
     and_i_choose_a_regular_confidence_level
     and_i_switch_between_evidence_type_to_other('Passport or travel document', 'UK passport', 'Something else')
+    and_i_say_the_evidence_strength_is('3')
     and_i_answer_no_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('3', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('Something else')
   end
 
@@ -38,7 +41,7 @@ RSpec.feature 'Add evidence flow', type: :system do
     and_i_choose_a_regular_confidence_level
     and_i_switch_between_evidence_type_to_regular('Something else', 'Passport or travel document', 'UK passport')
     and_i_answer_no_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('4', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('UK passport')
   end
 
@@ -47,7 +50,7 @@ RSpec.feature 'Add evidence flow', type: :system do
     and_i_choose_a_regular_confidence_level
     and_i_add_a_new_piece_of_evidence('Passport or travel document', 'UK passport')
     and_i_answer_yes_to_every_question
-    then_i_get_a_score('XX', 'out of 4')
+    then_i_get_a_score('4', 'out of 4')
     and_i_can_see_that_evidence_in_the_overview('UK passport')
   end
 
@@ -60,136 +63,43 @@ RSpec.feature 'Add evidence flow', type: :system do
     and_i_go_to_that_uri # skip submission on 'results' screen: evidence not marked 'complete'
     then_i_cannot_see_that_evidence_in_the_overview('UK passport')
   end
-end
 
-def and_i_add_a_new_piece_of_evidence(evidence_group, evidence_type)
-  click_link 'Add evidence'
-  choose evidence_group
-  choose evidence_type
-  click_button 'Continue'
-end
-
-def and_i_add_a_new_piece_of_other_evidence(evidence_other_name)
-  click_link 'Add evidence'
-  choose 'Other'
-  fill_in 'evidence_type_other', with: evidence_other_name
-  click_button 'Continue'
-end
-
-def and_i_answer_no_to_every_question
-  ['physical features', 'cryptographic security', 'authoritative source', 'cancelled'].each do |question_snippet|
-    expect(page).to have_content question_snippet
-    choose 'No'
-    click_button 'Continue'
-  end
-end
-
-def and_i_answer_yes_to_every_question
-  ['physical features', 'original', 'errors', 'document capture app', 'recognised guidance', 'taught how to tell'].each do |question_snippet|
-    expect(page).to have_content question_snippet
-    choose 'Yes'
+  def and_i_add_a_new_piece_of_other_evidence(evidence_other_name)
+    @evidence_text = evidence_other_name
+    click_link 'Add evidence'
+    choose 'Other'
+    fill_in 'evidence_type_other', with: evidence_other_name
     click_button 'Continue'
   end
 
-  expect(page).to have_content 'refresh their training'
-  choose 'Every year'
-  click_button 'Continue'
-
-  expect(page).to have_content 'official templates'
-  choose 'Yes'
-  click_button 'Continue'
-
-  expect(page).to have_content 'templates updated'
-  choose 'Every year'
-  click_button 'Continue'
-
-  ['expired', 'intercepted', 'visible security features', 'protects them from being tampered with'].each do |question_snippet|
-    expect(page).to have_content question_snippet
-    choose 'Yes'
+  def and_i_say_the_evidence_strength_is(strength)
+    expect(page).to have_content 'strength of this evidence'
+    choose strength.to_s
     click_button 'Continue'
   end
 
-  expect(page).to have_content 'security features'
-  check 'Designs printed using intaglio (raised) ink'
-  check 'Designs that have been laser etched'
-  check 'Watermarks'
-  check 'Security fibres'
-  check "Secondary background ('ghost') images"
-  click_button 'Continue'
+  def then_i_get_a_score_for_that_evidence(strength, strength_additional_text)
+    expect(page).to have_content "strength score of #{strength} #{strength_additional_text}"
+    click_button 'Continue'
+    find('th', text: @evidence_text).find(:xpath, "../../../tbody/tr[1]/td[2]").to have_content strength.to_s
+    # expect(find('???')).to have_content 'Change' # TODO
+  end
 
-  expect(page).to have_content 'consistent throughout'
-  choose 'Yes'
-  click_button 'Continue'
-
-  expect(page).to have_content 'equipment'
-  check 'magnification tool'
-  check 'low angle'
-  check 'Another piece of equipment'
-  check 'A non-specialist light source'
-  click_button 'Continue'
-
-  ['controlled', 'supervised', 'ultraviolet (UV) or infrared (IR)'].each do |question_snippet|
-    expect(page).to have_content question_snippet
-    choose 'Yes'
+  def and_i_switch_between_evidence_type_to_other(evidence_group, evidence_type, evidence_other_name)
+    click_link 'Add evidence'
+    choose evidence_group
+    choose evidence_type
+    choose 'Other'
+    fill_in 'evidence_type_other', with: evidence_other_name
     click_button 'Continue'
   end
 
-  expect(page).to have_content 'UV or IR security features'
-  check 'The paper the document is printed on'
-  check 'The layout and design of the document'
-  check 'Any fluorescent features'
-  click_button 'Continue'
-
-  expect(page).to have_content 'cryptographic security'
-  choose 'Yes'
-  click_button 'Continue'
-
-  expect(page).to have_content 'cryptographic security features'
-  check 'Check the evidence has not expired'
-  check 'Check the digital signature is correct'
-  check 'Check the signing key belongs to the organisation that issued the evidence'
-  check 'Check the signing key is correct for that type of evidence'
-  check 'Check the signing key has not been revoked'
-  click_button 'Continue'
-
-  ['authoritative source', 'cancelled'].each do |question_snippet|
-    expect(page).to have_content question_snippet
-    choose 'Yes'
+  def and_i_switch_between_evidence_type_to_regular(evidence_other_name, evidence_group, evidence_type)
+    click_link 'Add evidence'
+    choose 'Other'
+    fill_in 'evidence_type_other', with: evidence_other_name
+    choose evidence_group
+    choose evidence_type
     click_button 'Continue'
   end
-end
-
-def then_i_get_a_score(score, additional_text)
-  expect(page).to have_content "#{score} #{additional_text}"
-  click_button 'Continue'
-  # expect(find('??')).to have_text score # TODO
-  # expect(find('???')).to have_content 'Change' # TODO
-end
-
-def and_i_can_see_that_evidence_in_the_overview(evidence_type)
-  then_i_see_the_overview_screen
-  expect(page).to have_content evidence_type
-end
-
-def then_i_cannot_see_that_evidence_in_the_overview(evidence_type)
-  then_i_see_the_overview_screen
-  expect(page).not_to have_content evidence_type
-end
-
-def and_i_switch_between_evidence_type_to_other(evidence_group, evidence_type, evidence_other_name)
-  click_link 'Add evidence'
-  choose evidence_group
-  choose evidence_type
-  choose 'Other'
-  fill_in 'evidence_type_other', with: evidence_other_name
-  click_button 'Continue'
-end
-
-def and_i_switch_between_evidence_type_to_regular(evidence_other_name, evidence_group, evidence_type)
-  click_link 'Add evidence'
-  choose 'Other'
-  fill_in 'evidence_type_other', with: evidence_other_name
-  choose evidence_group
-  choose evidence_type
-  click_button 'Continue'
 end
